Fix joinMany (#56)
Instead of allocating an array of pointers, joinMany was allocating
memory for just one pointer. This was making ArrayFire read out of
bounds and fail with various errors.

This commit fixes this issue by adding a helper withManyForeignPtr
function that acts like withForeignPtr (not unsafeWithForeignPtr!), but
for a list of ForeignPtrs.