Fix (**) and use property tests (#57)
The default implementation of (**) relied on log and incorrectly
handled some inputs. The fix is making an explicit implementation
using the pow function.

To test the changes, tests for functions from the Floating typeclass
are re-written using property tests. There are a few helper functions
to make writing the actual properties easy.

More tests can be converted to properties, but this is left for
another PR.